added new basic api function for admin

+++ src/App/Repositories/Eloquent/UserRepository.php
@@ -2,9 +2,14 @@
 
 namespace App\Repositories\Eloquent;
 
+use App\Constants\RoleConstant;
+use App\Dtos\UserDto;
 use App\Models\User;
 use App\Repositories\Contracts\UserRepositoryInterface;
 use Billyranario\ProstarterKit\App\Helpers\LoggerHelper;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Contracts\Pagination\LengthAwarePaginator;
+use Illuminate\Log\Logger;
 use Illuminate\Support\Facades\{
     DB,
     Hash
@@ -13,7 +18,7 @@ use Illuminate\Support\Facades\{
 class UserRepository implements UserRepositoryInterface
 {
     /**
-     * @var User $model
+     * @var User $user
      */
     protected User $user;
 
@@ -26,6 +31,33 @@ class UserRepository implements UserRepositoryInterface
     {
         $this->user = $user;
     }
+
+    /**
+     * Get paginated users.
+     * @param UserDto $userDto
+     * @return LengthAwarePaginator
+     */
+    public function paginate(UserDto $userDto): LengthAwarePaginator
+    {
+        $keyword = $userDto->getSearchKeyword();
+
+        return $this->user
+            ->with($userDto->getRelations())
+            ->when($userDto->getIsArchive(), function (Builder $query) {
+                return $query->onlyTrashed();
+            })
+            ->when(!empty($keyword), function (Builder $user) use ($keyword) {
+                $user->where(function (Builder $query) use ($keyword) {
+                    $query->where('firstname', 'like', "%$keyword%")
+                        ->orWhere('lastname', 'like', "%$keyword%")
+                        ->orWhere('email', 'like', "%$keyword%");
+                });
+            })
+            ->where('role_id', '!=', RoleConstant::SUPERADMIN)
+            ->orderBy($userDto->getOrderBy(), $userDto->getOrderDirection())
+            ->paginate($userDto->getPerPage());
+    }
+
 
     /**
      * Find user by id.
@@ -99,6 +131,49 @@ class UserRepository implements UserRepositoryInterface
     }
 
     /**
+     * Bulk update users.
+     * @param array $data
+     * @param array $ids
+     * @return bool
+     */
+    public function bulkUpdate(array $data, array $ids): bool
+    {
+        try {
+            DB::beginTransaction();
+
+            $user = $this->user->whereIn('id', $ids)->update($data);
+
+            DB::commit();
+            return $user;
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            LoggerHelper::logThrowError($th);
+            return false;
+        }
+    }
+
+    /**
+     * Bulk restore users.
+     * @param array $ids
+     * @return bool
+     */
+    public function bulkRestore(array $ids): bool
+    {
+        try {
+            DB::beginTransaction();
+
+            $user = $this->user->whereIn('id', $ids)->restore();
+
+            DB::commit();
+            return $user;
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            LoggerHelper::logThrowError($th);
+            return false;
+        }
+    }
+
+    /**
      * Change Password
      * @param string $newPassword
      * @param int $id
@@ -152,11 +227,53 @@ class UserRepository implements UserRepositoryInterface
     /**
      * Delete user.
      * @param int $id
-     * @return bool|null
+     * @param bool $forceDelete
+     * @return bool
      */
-    public function delete(int $id): ?bool
+    public function delete(int $id, bool $forceDelete = false): bool
     {
         $user = $this->findById($id);
-        return $user->delete();
+
+        if ($user) {
+            if ($forceDelete) {
+                return $user->forceDelete();
+            }
+            return $user->delete();
+        }
+        return false;
+    }
+
+    /**
+     * Bulk delete users.
+     * @param array $ids
+     * @param bool $forceDelete
+     * @return bool
+     */
+    public function bulkDelete(array $ids, bool $forceDelete = false): bool
+    {
+        $query = $this->user->whereIn('id', $ids);
+
+        if ($forceDelete) {
+            $deletedCount = $query->forceDelete();
+        } else {
+            $deletedCount = $query->delete();
+        }
+
+        return $deletedCount > 0;
+    }
+
+    /**
+     * Restore archived user.
+     * @param int $id
+     * @return bool
+     */
+    public function restore(int $id): bool
+    {
+        $user = $this->findById($id);
+
+        if ($user) {
+            return $user->restore();
+        }
+        return false;
     }
 }
